refactor(step5): MoveStrategy domain 패키지로 이동
