+++ src/main/kotlin/study/racingcar/domain/MoveStrategy.kt
@@ -1,4 +1,4 @@
-package study.racingcar
+package study.racingcar.domain
 
 interface MoveStrategy {
     fun isMoveCar(): Boolean
